server/requirements.txt
# Core web framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0

# Gemini AI
google-generativeai==0.3.2

# Machine Learning et NLP
torch==2.1.0
sentence-transformers==2.2.2
transformers==4.35.0
spacy==3.7.2
scikit-learn==1.3.2
numpy==1.24.3

# Vector databases et recherche
faiss-cpu==1.7.4
pinecone-client==2.2.4

# Document processing
pypdf2==3.0.1
python-docx==1.0.1
beautifulsoup4==4.12.2
lxml==4.9.3
requests==2.31.0
aiohttp==3.9.0

# Database et cache
redis==5.0.1
psycopg2-binary==2.9.7

# Background tasks
celery==5.3.4

# Monitoring et logging
loguru==0.7.2

# French language model pour spaCy
# À installer séparément: python -m spacy download fr_core_news_lg

# Utilities
python-dateutil==2.8.2
python-multipart==0.0.6
python-jose[cryptography]==3.3.0